refactored carousels into 1 file
