+++ src/web/src/Components.ts
@@ -19,9 +19,5 @@ export { default as SideNavWrapper } from './Components/sidebar/SideNavWrapper';
 export { default as Searchbar } from './Components/Searchbar';
 export { default as RequireAuth } from './Components/auth/RequireAuth';
 export { default as Dropzone } from './Components/Dropzone';
-export { default as CS2Carousel } from './Components/CS2Carousel';
-export { default as ValorantCarousel } from './Components/ValorantCarousel';
+export { default as Carousel } from './Components/Carousel';
 export { default as Loading } from './Components/Loading';
-//Constants
-export const GAMES = ['CS2', 'Valorant'];
-export const CDN_URL = 'https://res.cloudinary.com/ddwqqjmyo/image/upload';
