+++ src/web/src/Constants.ts
@@ -1,0 +1,55 @@
+// VALORANT MAPS
+import ASCENT from './assets/valorantmaps/ascent.webp';
+import BIND from './assets/valorantmaps/bind.webp';
+import BREEZE from './assets/valorantmaps/breeze.webp';
+import FRACTURE from './assets/valorantmaps/fracture.webp';
+import HAVEN from './assets/valorantmaps/haven.webp';
+import ICEBOX from './assets/valorantmaps/icebox.webp';
+import LOTUS from './assets/valorantmaps/lotus.webp';
+import PEARL from './assets/valorantmaps/pearl.webp';
+import SPLIT from './assets/valorantmaps/split.webp';
+import SUNSET from './assets/valorantmaps/sunset.webp';
+import valorantbanner from './assets/valorantbanner.webp';
+
+// CS2 MAPS
+import ANCIENT from './assets/cs2maps/ancient.webp';
+import ANUBIS from './assets/cs2maps/anubis.webp';
+import DUST2 from './assets/cs2maps/dust2.webp';
+import INFERNO from './assets/cs2maps/inferno.webp';
+import MIRAGE from './assets/cs2maps/mirage.webp';
+import NUKE from './assets/cs2maps/nuke.webp';
+import OVERPASS from './assets/cs2maps/overpass.webp';
+import VERTIGO from './assets/cs2maps/vertigo.webp';
+import cs2banner from './assets/cs2banner.png';
+
+export const VALORANT_MAPS = [
+   ASCENT,
+   BIND,
+   BREEZE,
+   FRACTURE,
+   HAVEN,
+   ICEBOX,
+   LOTUS,
+   PEARL,
+   SPLIT,
+   SUNSET,
+];
+
+export const VALORANT_BANNER = valorantbanner;
+
+export const CS2_MAPS = [
+   ANCIENT,
+   ANUBIS,
+   DUST2,
+   INFERNO,
+   MIRAGE,
+   NUKE,
+   OVERPASS,
+   VERTIGO,
+];
+
+export const CS2_BANNER = cs2banner;
+
+export const GAMES = ['CS2', 'Valorant'];
+
+export const CDN_URL = 'https://res.cloudinary.com/ddwqqjmyo/image/upload';
